+++ src/core/CanKit.Core/Registry/CanRegistry.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Threading;
 using CanKit.Abstractions.API.Can;
 using CanKit.Abstractions.API.Common;
@@ -213,64 +211,17 @@ public partial class CanRegistry
 
     internal static CanRegistry? Instance;
 
-    internal CanRegistry(params Assembly[] assembliesToScan)
+    internal CanRegistry()
     {
         Instance = this;
-        var assemblies = assembliesToScan.Length == 0 ? [Assembly.GetExecutingAssembly()] : assembliesToScan;
-        ExecuteRegistrationPipeline(assemblies);
     }
 
-    private static Assembly Entry()
-    {
-
-        var entry = Assembly.GetEntryAssembly();
-        if (entry != null) return entry;
-
-        var withEntry = AppDomain.CurrentDomain.GetAssemblies()
-            .FirstOrDefault(a => a.EntryPoint != null);
-        if (withEntry != null) return withEntry;
-
-        return Assembly.GetExecutingAssembly();
-    }
     private static CanRegistry BuildRegistry()
     {
-        try
-        {
-            var genType = Entry().GetType("CanKit.Core.Internal.AdapterPreloadList", false);
-
-            if (genType?
-                    .GetField("Assemblies", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static)?
-                    .GetValue(null) is string[] names)
-            {
-                foreach (var n in names.Where(s => !string.IsNullOrWhiteSpace(s)))
-                {
-                    try { SafeLoad(new AssemblyName(n)); }
-                    catch { /* ignore one-off load errors */ }
-                }
-            }
-        }
-        catch { /* ignore any preload reflection errors */ }
-        var asms = AppDomain.CurrentDomain.GetAssemblies()
-            .Where(a => !a.IsDynamic)
-            .ToArray();
-
-        var reg = new CanRegistry(asms);
-
-        return reg;
-    }
-
-    private static void SafeLoad(AssemblyName path)
-    {
-        try
-        {
-#if NET5_0_OR_GREATER
-            var asm = System.Runtime.Loader.AssemblyLoadContext.Default.LoadFromAssemblyName(path);
-#else
-            var asm = Assembly.Load(path); // .NET Framework
-#endif
-
-        }
-        catch {/* ignore any preload reflection errors */ }
+        var snapshot = CanRegistryBootstrap.Freeze();
+        var registry = new CanRegistry();
+        registry.ExecuteRegistrationPipeline(snapshot);
+        return registry;
     }
 }