+++ tests/CanKit.Tests/Utils/IsExternalInit.cs
@@ -1,4 +1,3 @@
-using System;
 using System.ComponentModel;
 
 #if !NET5_0_OR_GREATER
@@ -6,8 +5,5 @@ namespace System.Runtime.CompilerServices
 {
     [EditorBrowsable(EditorBrowsableState.Never)]
     internal static class IsExternalInit;
-
-    [AttributeUsage(AttributeTargets.Method, Inherited = false)]
-    internal sealed class ModuleInitializerAttribute : Attribute;
 }
 #endif